confirmation for deletion of reminders

+++ Application/SwinApp/SwinApp/SwinApp/Components/UI/CardReminder.xaml.cs
@@ -27,8 +27,11 @@ namespace SwinApp.Components
             ButtonDeleteReminder.GestureRecognizers.Add(tap);
         }
 
-        public void DeleteReminder()
+        public async void DeleteReminder()
         {
+
+            var confirmDeletion = await App.Current.MainPage.DisplayAlert("Delete?", "Delete this reminder?", "Yes", "No");
+            if (confirmDeletion)
             User.DeleteReminder(_reminder);
         }
